+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/feature/MessageShowcaseActivity.kt
@@ -114,12 +114,10 @@ class MessageShowcaseActivity : AppCompatActivity() {
             val changeMessage = layoutMessagesChange.findViewById<AndesMessage>(R.id.message)
 
             val links = listOf(
-                    AndesBodyLink(0, 5),
+                    AndesBodyLink(4, 11),
+                    AndesBodyLink(60, 66),
                     AndesBodyLink(79, 122),
-                    AndesBodyLink(50, 40),
-                    AndesBodyLink(79, 124),
-                    AndesBodyLink(-1, 10),
-                    AndesBodyLink(-1, -10)
+                    AndesBodyLink(50, 40)
             )
 
             changeMessage.bodyLinks = (AndesBodyLinks(links, listener = {
@@ -142,6 +140,7 @@ class MessageShowcaseActivity : AppCompatActivity() {
                 changeMessage.title = titleText.text.toString()
                 changeMessage.type = AndesMessageType.fromString(typeSpinner.selectedItem.toString())
                 changeMessage.hierarchy = AndesMessageHierarchy.fromString(hierarchySpinner.selectedItem.toString())
+                changeMessage.bodyLinks = null
 
                 if (primaryActionText.text.toString().isNotEmpty()) {
                     changeMessage.setupPrimaryAction(primaryActionText.text.toString(), View.OnClickListener {
@@ -251,6 +250,15 @@ class MessageShowcaseActivity : AppCompatActivity() {
                         Toast.makeText(context, "Link onClick", Toast.LENGTH_SHORT).show()
                     }
             )
+            val links = listOf(
+                    AndesBodyLink(6, 11),
+                    AndesBodyLink(64, 71)
+            )
+            layoutMessages.findViewById<AndesMessage>(R.id.messageLinkBody).bodyLinks = (AndesBodyLinks(links,
+                listener = {
+                    Toast.makeText(context, "Click at body link: $it", Toast.LENGTH_SHORT).show()
+                }
+            ))
             return layoutMessages
         }
     }