+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v2.7.0
 ## Added
 - LinkButton in AndesMessage Component.
+- Thumbnail Component, Icon type.
 
 # v2.6.0
 ## Added